refactor(core): Move tab-specific code out of core
While the two UIs share similar functionality, it has been in distinct methods all along. It should be refactored into common parts later.
